Rewrite app-service-registration in Typescript
